Implement packages separately in jvm and js
