+++ js/src/main/scala/mouse/js.scala
@@ -1,0 +1,3 @@
+package mouse
+
+trait AllJsSyntax
