+++ jvm/src/main/scala/mouse/jvm.scala
@@ -1,0 +1,3 @@
+package mouse
+
+trait AllJvmSyntax
